move prefixes into types
